Backend routes and controllers
API integration is in progress. Right now the structure for the routes and controllers and how they connect with each other, and the controllers to the services is implemented. Right now all the routes for the main aspect of our project is established. The server file is also established. I have also initialized prisma and have a prisma.schema with an unfinished model. Still waiting to migrate.