Backend routes, controllers, and services integration with Weather APi and db migration
The backend server is up and running on Port 3000. Weather APIs and azure sql DB works with the backend server. The schema is migrated. All the routes, controllers, and services for the main part of the app are coded. A heap is implemented to sort and display the people in danger by priority. All GET and POST endpoints are implemented and connected with backend database. Still needs integration and testing with frontend.